Day 2, part 2
I had to get some help with this one. After failing to do a similiar
approach as for part 1, only expanding the window to three levels, I was
thinking to just take all combinations of a report where I leave out a
level. The problem was I couldn't come up with a function for achieving
this. I understand how it works, but I still don't think I could have
come up with it. More practice!