Add a coordinate utility
The main problem with this implementation is that I don't really know
how I got it to work. There was a lot of trial and error to get it
right, and in the end I don't really know what the trick was to get it
working. Pulling in the core module probably didn't make things easier
to follow.